feat: add engine init endpoint
