+++ cortex-js/src/domain/config/config.interface.ts
@@ -1,6 +1,5 @@
 export interface Config {
   dataFolderPath: string;
-  initialized: boolean;
   cortexCppHost: string;
   cortexCppPort: number;
 }
