+++ cortex-js/src/infrastructure/commanders/test/helpers.command.spec.ts
@@ -35,7 +35,6 @@ beforeAll(
       // Attempt to create test folder
       await fileService.writeConfigFile({
         dataFolderPath: join(__dirname, 'test_data'),
-        initialized: false,
         cortexCppHost: 'localhost',
         cortexCppPort: 3929
       });
