+++ cortex-js/src/infrastructure/commanders/test/models.command.spec.ts
@@ -23,7 +23,6 @@ beforeAll(
       // Attempt to create test folder
       await fileService.writeConfigFile({
         dataFolderPath: join(__dirname, 'test_data'),
-        initialized: false,
         cortexCppHost: 'localhost',
         cortexCppPort: 3929,
       });
